* Headers/AppKit/NSTextView.h: Change _dragTargetLocation to NSUInteger because NSNotFound no longer fits in unsigned int on 64-bit
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@34295 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSTextView.h
@@ -230,7 +230,7 @@ therefore be stored in the NSLayoutManager to avoid problems.
   /* Ivar to store the location where text is going to be inserted during
    * a DnD operation.
    */
-  unsigned int _dragTargetLocation;
+  NSUInteger _dragTargetLocation;
 
   /* Ivar used to implement coalescing of undo operations */
   id _undoObject;
